modified:   source/fixes/EliteForcesWW2IwoJimaFOVFix/dllmain.cpp 	modified:   source/fixes/KISSPsychoCircusTheNightmareChildFOVFix/dllmain.cpp

+++ source/fixes/EliteForcesWW2IwoJimaFOVFix/dllmain.cpp
@@ -55,7 +55,7 @@ float fFOVFactor;
 float fAspectRatioScale;
 float fNewCameraHFOV;
 float fNewCameraVFOV;
-static uint32_t iIsUnderwater;
+static uint16_t iIsUnderwater;
 
 // Game detection
 enum class Game
@@ -254,7 +254,7 @@ void FOVFix()
 
 			UnderwaterValueCheckMidHook = safetyhook::create_mid(UnderwaterValueCheckInstructionScanResult, [](SafetyHookContext& ctx)
 			{
-				iIsUnderwater = *reinterpret_cast<uint32_t*>(ctx.esi + 0x61);
+				iIsUnderwater = *reinterpret_cast<uint16_t*>(ctx.esi + 0x61);
 			});
 		}
 		else
